Run scalafmt
